Empty object log fix

+++ src/js/03-feedback.js
@@ -1,36 +1,47 @@
-import throttle from "lodash.throttle";
+import throttle from 'lodash.throttle';
 
-const form = document.querySelector(".feedback-form");
-const emailInput = form.querySelector("[name = email]");
-const messageInput = form.querySelector("[name = message]");
+const form = document.querySelector('.feedback-form');
+const emailInput = form.querySelector('[name = email]');
+const messageInput = form.querySelector('[name = message]');
+const formStateKey = 'feedback-form-state';
 const feedback = {};
-const formStateKey = "feedback-form-state";
-const formState = localStorage.getItem(formStateKey);
 
-form.addEventListener("input", throttle(saveFormValues, 500));
-form.addEventListener("submit", onFormSubmit);
+getValues();
 
-if (formState) {
-  try {
-    const feedbackValues = JSON.parse(formState);
+form.addEventListener('input', throttle(saveFormValues, 500));
+form.addEventListener('submit', onFormSubmit);
 
-    emailInput.value = feedbackValues.email;
-    messageInput.value = feedbackValues.message;
-  } catch {
-    console.log("Невалідні дані");
+function getValues() {
+  if (localStorage.getItem(formStateKey)) {
+    try {
+      const feedbackValues = JSON.parse(localStorage.getItem(formStateKey));
+
+      emailInput.value = feedbackValues.email;
+      messageInput.value = feedbackValues.message;
+
+      feedback.email = feedbackValues.email;
+      feedback.message = feedbackValues.message;
+    } catch {
+      console.log('Невалідні дані');
+    }
   }
 }
 
 function saveFormValues() {
-  feedback.email = emailInput.value;
-  feedback.message = messageInput.value;
-
-  localStorage.setItem(formStateKey, JSON.stringify(feedback));
+  if (emailInput.value || messageInput.value) {
+    localStorage.setItem(formStateKey, JSON.stringify(feedback));
+    feedback.email = emailInput.value;
+    feedback.message = messageInput.value;
+  } else {
+    localStorage.clear();
+  }
 }
 
 function onFormSubmit(e) {
   e.preventDefault();
-  form.reset();
-  localStorage.clear();
-  console.log(feedback);
+  if (emailInput.value || messageInput.value) {
+    console.log(feedback);
+    form.reset();
+    localStorage.clear();
+  }
 }
